superadmin update

+++ app/accounts/deshboard/schema.py
@@ -5,8 +5,8 @@ class TopClientOut(BaseModel):
     client_id: int
     name: str
     total_orders: int
-    revenue: int
+    revenue: float
     growth: float | None = 0
-
+ 
     class Config:
         from_attributes = True
